Create index.html
As you know login is one of most priorities for your website users. So here we design a login form for you with beautiful CSS with a background image. In this template, we use box-shadow CSS properties for the show background image on login form like transparent and we used a bootstrap CSS framework to make this login form with given classes in this framework.